requirements.txt
telethon>=1.28.5
python-dotenv>=0.19.0
transformers>=4.30.0
python-decouple>=3.8
torch>=2.0.0  # Для NLP моделей